requirements.txt
# requirements.txt

accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bcrypt==5.0.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
docling==2.46.0
docling-core==2.45.0
docling-ibm-models==3.9.0
docling-parse==4.2.2
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-docling==1.0.0
langchain-openai==0.3.31
langchain-pinecone==0.2.11
langchain-tavily==0.2.11
langchain-tests==0.3.20
langchain-text-splitters==0.3.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langgraph-supervisor==0.0.29
langsmith==0.4.15
latex2mathml==3.78.0
lazy_loader==0.4
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
networkx==3.5
ninja==1.13.0
numpy==2.3.2
openai==1.101.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.11.2
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.2
passlib==1.7.4
pillow==11.3.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pylatexenc==2.10
pypdfium2==4.30.0
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==4.0.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
starlette==0.47.2
sympy==1.14.0
syrupy==4.9.1
tabulate==0.9.0
tenacity==9.1.2
tifffile==2025.6.11
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.3
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vcrpy==7.0.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
